requirements.txt
bip-utils
mnemonic
ecdsa
requests
bitcoinlib
supabase 
python-dotenv